+++ docs/source/usage/examples.rst
@@ -83,6 +83,7 @@ Channels & Rings
    :maxdepth: 1
 
    examples/fodo_channel/README.rst
+   examples/solenoid_restart/README.rst
 
 
 Lattice Design & Optimization
