+++ src/elements/CMakeLists.txt
@@ -2,6 +2,7 @@ target_sources(lib
   PRIVATE
     Aperture.cpp
     Programmable.cpp
+    Source.cpp
 )
 
 add_subdirectory(diagnostics)
